Refactoring Products component architechture
