+++ src/Products/List/Grid/Card/styled.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import { LabelLayout } from '../../../components/SaleLabel';
+import { LabelLayout } from '../../../../components/SaleLabel';
 
 export const Image = styled.img`
   width: 100%;
