+++ src/Products/Show/Meta/Right/ColorSwitcher/index.jsx
