+++ src/Products/Show/Meta/Right/ColorSwitcher/styled.js
